[issue-381] change error type of datetime_from_str()
