+++ src/datetime_conversions.py
@@ -15,12 +15,9 @@ from src.parser.error import SPDXParsingError
 
 def datetime_from_str(date_str: str) -> datetime:
     if not isinstance(date_str, str):
-        raise SPDXParsingError([f"Could not convert str to datetime, invalid type: {type(date_str).__name__}"])
-    try:
-        date = datetime.strptime(date_str, "%Y-%m-%dT%H:%M:%SZ")
-    except ValueError:
-        raise SPDXParsingError(
-            [f'Could not convert str to datetime, format of {date_str} does not match "%Y-%m-%dT%H:%M:%SZ"'])
+        raise TypeError(f"Could not convert str to datetime, invalid type: {type(date_str).__name__}")
+
+    date = datetime.strptime(date_str, "%Y-%m-%dT%H:%M:%SZ") # raises ValueError if format does not match
     return date
 
 def datetime_to_iso_string(date: datetime) -> str:
