+++ tests/test_datetime_conversions.py
@@ -1,0 +1,31 @@
+#  Copyright (c) 2022 spdx contributors
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at
+#      http://www.apache.org/licenses/LICENSE-2.0
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+#  limitations under the License.
+from datetime import datetime
+
+import pytest
+
+from src.datetime_conversions import datetime_from_str
+
+
+def test_datetime_from_str():
+    date_str = "2010-03-04T05:45:11Z"
+
+    date = datetime_from_str(date_str)
+
+    assert date == datetime(2010, 3, 4, 5, 45, 11)
+
+
+@pytest.mark.parametrize("invalid_date_str, error_type, expected_message",
+                         [(5, TypeError, "Could not convert str to datetime, invalid type: int"),
+                          ("2010-02-03", ValueError, "time data '2010-02-03' does not match format '%Y-%m-%dT%H:%M:%SZ'")])
+def test_datetime_from_str_error(invalid_date_str, error_type, expected_message):
+    with pytest.raises(error_type, match=expected_message):
+        datetime_from_str(invalid_date_str)
